Crate sc_fast_sync_backend

Fast-Sync Backend

This is modified from the Substrate in memory backend, designated solely for executing the blocks efficiently in the memory during the initial sync. This backend only maintains the latest state of chain rather than the state history.

Structs

Backend

In-memory backend. Keeps the latest state in memory.

BlockImportOperation

In-memory operation.

Blockchain

In-memory blockchain. Supports concurrent reads.

Functions

check_genesis_storage

Check that genesis storage is valid.