Struct sc_fast_sync_backend::Backend

pub struct Backend<Block: BlockT> { /* private fields */ }

In-memory backend. Keeps the latest state in memory.

Warning: Doesn’t support all the features necessary for a proper database. Only use this struct for the fast sync purposes. Do NOT use in production.

Implementations

impl<Block: BlockT> Backend<Block>

pub fn new(runtime_hash: Block::Hash, runtime_code: Vec<u8>) -> Self

Create a new instance of in-mem backend for the fast sync purpose.

pub fn initialize( &mut self, best_header: Block::Header, genesis_hash: Block::Hash, state: InMemoryBackend<HashingFor<Block>>, )

Initialize the blockchain storage and chain state.

pub fn storage_root(&self) -> Block::Hash

pub fn pin_refs(&self, hash: &<Block as BlockT>::Hash) -> Option<i64>

Return the number of references active for a pinned block.

Warning

For testing purposes only!